starterkit/requirements.txt
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
requests>=2.31.0
jinja2>=3.1.0
